+++ src/any_value/lazy_clone.rs
@@ -1,5 +1,6 @@
 use std::any::TypeId;
-use crate::any_value::{AnyValue, AnyValueCloneable, AnyValueUnknown};
+use std::ptr::NonNull;
+use crate::any_value::{AnyValueTyped, AnyValueCloneable, AnyValueSized, AnyValuePtr};
 
 /// Makes [`AnyValueCloneable`] actually [`Clone`]able.
 /// Do clone on consumption.
@@ -20,21 +21,28 @@ impl<'a, T: AnyValueCloneable> LazyClone<'a, T>{
     }
 }
 
-impl<'a, T: AnyValueCloneable> AnyValueUnknown for LazyClone<'a, T>{
+impl<'a, T: AnyValueCloneable> AnyValuePtr for LazyClone<'a, T> {
     type Type = T::Type;
 
     #[inline]
-    fn as_bytes(&self) -> &[u8]{
-        self.value.as_bytes()
+    fn as_bytes_ptr(&self) -> NonNull<u8> {
+        self.value.as_bytes_ptr()
     }
 
     #[inline]
-    unsafe fn move_into(self, out: *mut u8) {
+    unsafe fn move_into<KnownType:'static /*= Unknown*/>(self, out: *mut u8, _bytes_size: usize) {
         self.value.clone_into(out);
     }
 }
 
-impl<'a, T: AnyValueCloneable + AnyValue> AnyValue for LazyClone<'a, T>{
+impl<'a, T: AnyValueCloneable + AnyValueSized> AnyValueSized for LazyClone<'a, T>{
+    #[inline]
+    fn size(&self) -> usize {
+        self.value.size()
+    }
+}
+
+impl<'a, T: AnyValueCloneable + AnyValueTyped> AnyValueTyped for LazyClone<'a, T>{
     #[inline]
     fn value_typeid(&self) -> TypeId {
         self.value.value_typeid()