+++ src/ops/splice.rs
@@ -1,11 +1,11 @@
 use crate::any_vec_ptr::IAnyVecRawPtr;
 use crate::{any_vec_ptr, assert_types_equal, Iter};
-use crate::any_value::{AnyValueUnknown, AnyValue};
+use crate::any_value::{AnyValueTyped, move_into_w_size};
 use crate::ops::iter::Iterable;
 
 pub struct Splice<'a, AnyVecPtr: IAnyVecRawPtr, ReplaceIter: ExactSizeIterator>
 where
-    ReplaceIter::Item: AnyValue
+    ReplaceIter::Item: AnyValueTyped
 {
     iter: Iter<'a, AnyVecPtr>,
     start: usize,
@@ -16,7 +16,7 @@ where
 impl<'a, AnyVecPtr: IAnyVecRawPtr, ReplaceIter: ExactSizeIterator>
     Splice<'a, AnyVecPtr, ReplaceIter>
 where
-    ReplaceIter::Item: AnyValue
+    ReplaceIter::Item: AnyValueTyped
 {
     #[inline]
     pub fn new(
@@ -44,7 +44,7 @@ impl<'a, AnyVecPtr: IAnyVecRawPtr, ReplaceIter: ExactSizeIterator> Iterable
 for
     Splice<'a, AnyVecPtr, ReplaceIter>
 where
-    ReplaceIter::Item: AnyValue
+    ReplaceIter::Item: AnyValueTyped
 {
     type Iter = Iter<'a, AnyVecPtr>;
 
@@ -63,7 +63,7 @@ impl<'a, AnyVecPtr: IAnyVecRawPtr, ReplaceIter: ExactSizeIterator> Drop
 for
     Splice<'a, AnyVecPtr, ReplaceIter>
 where
-    ReplaceIter::Item: AnyValue
+    ReplaceIter::Item: AnyValueTyped
 {
     fn drop(&mut self) {
         use any_vec_ptr::utils::*;
@@ -105,7 +105,7 @@ where
             let mut ptr = element_mut_ptr_at(any_vec_ptr, self.start);
             while let Some(replace_element) = self.replace_with.next() {
                 assert_types_equal(type_id, replace_element.value_typeid());
-                replace_element.move_into(ptr);
+                move_into_w_size(replace_element, ptr, element_size);
                 ptr = ptr.add(element_size);
             }
         }